Move Edit Status to Ticket Component
I moved the HTML select element from the EditTicket route and moved it
to the Ticket route. In doing so, I created a axios.patch call to send
the new status. I added the patch handling to the backend and updated
the fake database's updateTicket to handle this partial ticket update.

[Ticket: ZTH-4]