Add Notifications to Ticket: Respond and Edit
I added a notification-container component to the `Ticket` component  as a parking
spot for notifications. Adjusting the state for editing the Status was
more straight-forward because the status selector was moved to the
component, itself. The `handleStatusChange` function sets the state that
allows for conditional rendering of the notification.

The Response notification was a little more tricky since the Response comes
from a different route and (currently) isn't implemented as a child
component that inherits props. What I settled on was adding a searchParam
`({responseSubmitted: "true"})` to the `Response` component's redirect. The
Ticket's useEffect, in turn, checks for this searchParam and, if true,
fires a memoized function `handleResponseSubmitted`. This function
adjusts the state of `responseSubmitted` and clears the searchParam from
the URL.

Overall, I'm happy with the performance of this workaround, but I think
the `Ticket` code isn't as clean. Open for suggestions! I wonder if I
could clear a lot of this up by making the `/respond/` path a child of
Ticket, to allow for prop handling.

[Ticket: ZTH-5 ]