Fix the New Ticket Button
Responding to the critique: I would argue that there is some value in
being able to create a New Ticket from the admin panel. Use case: a
customer submits a ticket through an unoficial means (i.e. emailing
hello@business.com, or as a suggestion from someone within the
company). I made the process more streamlined by going straight to the
Edit page upon clicking New Ticket.

To fit the existing architecture, clicking `New Ticket` creates a new
ticket {name: "New Ticket", email: "", description: ""} and then the
edit page updates the ticket's information or deletes it.

A more optimal implementation would be to make it so that the new ticket
is only created upon clicking Submit on the edit page. This suggestion
is listed as ticket ZTH-19.

[Ticket: ZTH-6]