Add Comments to Tickets
I added a comment section component that is added into the ticket. In
its current implemention, a user must first write their name to leave a
comment. Once auth is implemented, this can be replaced with the userId.
Further improvements: ability to edit or delete past comments written by
the user.

I decided to create a new directory, components. In this project, I am
differentiating between page components (under Routes) that are
connected to a specific URL paths and are responsible for fetching and
rendering content for that route. The components directory is for
reusable UI components that are not tied to the route. I'm open to other
ways of organizing this project!

[Ticket: ZTH-12]